+++ tools/integration_tests/stale_handle/stale_file_handle_open_file_test.go
@@ -1,0 +1,96 @@
+package stale_handle
+
+import (
+	"log"
+	"os"
+	"path"
+	"syscall"
+	"testing"
+
+	"cloud.google.com/go/storage"
+	"github.com/googlecloudplatform/gcsfuse/v3/internal/util"
+	. "github.com/googlecloudplatform/gcsfuse/v3/tools/integration_tests/util/client"
+	"github.com/googlecloudplatform/gcsfuse/v3/tools/integration_tests/util/operations"
+	"github.com/googlecloudplatform/gcsfuse/v3/tools/integration_tests/util/setup"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/suite"
+)
+
+type staleFileHandleOpenFile struct {
+	suite.Suite
+	flags                    []string
+	f1                       *os.File
+	fileName                 string
+	data                     string
+	isStreamingWritesEnabled bool
+	isLocal                  bool
+}
+
+func (s *staleFileHandleOpenFile) SetupSuite() {
+	setup.MountGCSFuseWithGivenMountWithConfigFunc(testEnv.cfg, s.flags, mountFunc)
+	setup.SetMntDir(mountDir)
+	testEnv.testDirPath = SetupTestDirectory(testEnv.ctx, testEnv.storageClient, testDirName)
+	s.data = setup.GenerateRandomString(5 * util.MiB)
+}
+
+func (s *staleFileHandleOpenFile) TearDownSuite() {
+	setup.UnmountGCSFuseWithConfig(testEnv.cfg)
+}
+
+func (s *staleFileHandleOpenFile) SetupTest() {
+	s.fileName = path.Base(s.T().Name()) + setup.GenerateRandomString(5)
+	err := CreateObjectOnGCS(testEnv.ctx, testEnv.storageClient, path.Join(testDirName, s.fileName), "")
+	assert.NoError(s.T(), err)
+	s.f1 = operations.OpenFileWithODirect(s.T(), path.Join(testEnv.testDirPath, s.fileName))
+}
+
+func (s *staleFileHandleOpenFile) TearDownTest() {
+	setup.SaveGCSFuseLogFileInCaseOfFailure(s.T())
+}
+
+func (s *staleFileHandleOpenFile) TestOpenFileWhenClobbered() {
+	defer func() { _ = s.f1.Close() }()
+
+	// 1. Get old inode ID.
+	fi1, err := os.Stat(s.f1.Name())
+	assert.NoError(s.T(), err)
+	stat1 := fi1.Sys().(*syscall.Stat_t)
+	oldInodeId := stat1.Ino
+
+	// 2. Clobber the file on GCS.
+	err = WriteToObject(testEnv.ctx, testEnv.storageClient, path.Join(testDirName, s.fileName), FileContents, storage.Conditions{})
+	assert.NoError(s.T(), err)
+
+	// 3. Open the file again. It should succeed (VFS retry).
+	fh2, err := os.OpenFile(path.Join(testEnv.testDirPath, s.fileName), os.O_RDWR|syscall.O_DIRECT, 0)
+	assert.NoError(s.T(), err)
+	defer operations.CloseFileShouldNotThrowError(s.T(), fh2)
+
+	// 4. Get new inode ID.
+	fi2, err := fh2.Stat()
+	assert.NoError(s.T(), err)
+	stat2 := fi2.Sys().(*syscall.Stat_t)
+	newInodeId := stat2.Ino
+
+	// 5. Verify they are different.
+	assert.NotEqual(s.T(), oldInodeId, newInodeId)
+
+	// 6. Verify we can write to the new fd.
+	_, err = fh2.Write([]byte("chips"))
+	assert.NoError(s.T(), err)
+}
+
+func TestStaleHandleOpenFileWhenClobbered(t *testing.T) {
+	if setup.AreBothMountedDirectoryAndTestBucketFlagsSet() {
+		suite.Run(t, new(staleFileHandleOpenFile))
+		return
+	}
+
+	flagsSet := setup.BuildFlagSets(*testEnv.cfg, testEnv.bucketType, t.Name())
+	for _, flags := range flagsSet {
+		s := new(staleFileHandleOpenFile)
+		s.flags = flags
+		log.Printf("Running OpenFile tests with flags: %s", s.flags)
+		suite.Run(t, s)
+	}
+}
